split up platfor-team policy packs
