--- platform-team/PulumiPolicy.yaml
@@ -1,2 +1,0 @@
-runtime: nodejs
-description: A minimal Policy Pack for Azure using TypeScript.
