+++ platform-team/approved-component-versions/.gitignore
