+++ platform-team/approved-component-versions/package.json
@@ -1,0 +1,11 @@
+{
+    "name": "approved-component-versions",
+    "version": "1.0.0",
+    "dependencies": {
+        "@pulumi/pulumi": "^3.0.0",
+        "@pulumi/policy": "^1.3.0"
+    },
+    "devDependencies": {
+        "@types/node": "^10.0.0"
+    }
+}
