+++ platform-team/approved-component-versions/tsconfig.json
