+++ platform-team/approved-components-resources/.gitignore
@@ -1,0 +1,2 @@
+/bin/
+/node_modules/
