+++ platform-team/approved-components-resources/approved-components-resources.json
@@ -1,0 +1,24 @@
+{
+  "approved-components-and-resource-types": {
+    "enforcementLevel": "advisory",
+    "approvedComponentTypes": [
+      "containerapps:index:AppBuildDeploy",
+      "containerapps:index:AppDeploy", 
+      "containerapps:index:AppImage", 
+      "container-services:index:AppDeploy", 
+      "container-services:index:AppImage", 
+      "container-services:index:AppImageDeploy", 
+      "cloudfront-s3-cdk:index:CloudFrontS3", 
+      "k8sapp:index:ServiceDeployment",
+      "stackmgmt:index:stacksettings", 
+      "pequod:gke:Cluster",
+      "awsx:lb:ApplicationLoadBalancer",
+      "awsx:ecs:FargateService",
+      "awsx:ecs:FargateTaskDefinition",
+      "awsx:ecr:Repository"
+    ],
+    "allowedResourceTypes": [
+      "azure-native:resources:ResourceGroup"
+    ]
+  }
+}
