+++ platform-team/approved-components-resources/package.json
@@ -1,6 +1,6 @@
 {
-    "name": "platform-team",
-    "version": "1.5.2",
+    "name": "approved-components-and-resource-types",
+    "version": "1.0.0",
     "dependencies": {
         "@pulumi/pulumi": "^3.0.0",
         "@pulumi/policy": "^1.3.0"
